WIP: 2024-04-26 21:06

+++ server/controllers/cardsController.js
@@ -1,13 +1,14 @@
 const Card = require('../models/Card.js');
-const User = require('../models/User.js');
-const Board = require('../models/Board.js');
 const List = require('../models/List.js');
 const mongoose = require('mongoose');
 
-const userByUsername = (username, option = { lean: true }) => {
-    const foundUser = User.findOne({ username });
-    if (option.lean) foundUser.lean();
-    return foundUser;
+const { isActionAuthorized } = require('../services/boardActionAuthorizeService');
+const { userByUsername } = require('../services/userService');
+
+const listById = (id, option = { lean: true }) => {
+    const foundList = List.findById(id);
+    if (option.lean) foundList.lean();
+    return foundList;
 };
 
 const cardById = (id, option = { lean: true }) => {
@@ -32,59 +33,108 @@ const getCard = async (req, res) => {
 const addCard = async (req, res) => {
     const { trackedId, title, order, listId } = req.body;
 
+    const foundList = await listById(listId);
+    if (!foundList) return res.status(403).json({ msg: "list not found" });
+
+    const { boardId } = foundList;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
     const newCard = new Card({
         trackedId,
         title,
         order,
-        listId
+        listId,
+        boardId,
     });
 
     await newCard.save();
     return res.status(201).json({ msg: 'new card added', newCard });
 };
 
-const updateCard = async (req, res) => {
-    const { id } = req.params;
-    const { title, description, order, listId } = req.body;
-    const newCard = await Card.findOneAndUpdate({ _id: id }, { title, description, order, listId }, { new: true });
-    return res.status(200).json({ msg: 'new card updated', newCard });
-};
-
 const reorder = async (req, res) => {
     const { id } = req.params;
     const { rank, listId } = req.body;
-    const newCard = await Card.findOneAndUpdate({ _id: id }, { order: rank, listId }, { new: true });
-    res.status(200).json({ message: 'card updated', newCard });
+
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
+
+    const foundList = await listById(listId);
+    if (!foundList) return res.status(404).json({ error: 'List not found' });
+
+    const { boardId } = foundList;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
+    foundCard.rank = rank;
+    foundCard.listId = listId;
+    await foundCard.save();
+
+    res.status(200).json({ message: 'card updated', newCard: foundCard });
 };
 
 const updateTitle = async (req, res) => {
     const { id } = req.params;
     const { title } = req.body;
-    const newCard = await Card.findOneAndUpdate({ _id: id }, { title }, { new: true });
-    res.status(200).json({ message: 'card updated', newCard });
+
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
+
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
+    foundCard.title = title;
+    await foundCard.save();
+
+    res.status(200).json({ message: 'card updated', newCard: foundCard });
 };
 
 const updateDescription = async (req, res) => {
     const { id } = req.params;
     const { description } = req.body;
-    const newCard = await Card.findOneAndUpdate({ _id: id }, { description }, { new: true });
-    res.status(200).json({ message: 'card updated', newCard });
+
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
+
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
+    foundCard.description = description;
+    await foundCard.save();
+
+    res.status(200).json({ message: 'card updated', newCard: foundCard });
 };
 
 const updateHighlight = async (req, res) => {
     const { id } = req.params;
     const { highlight } = req.body;
-    const newCard = await Card.findOneAndUpdate({ _id: id }, { highlight }, { new: true });
-    res.status(200).json({ message: 'card updated', newCard });
+
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
+
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
+    foundCard.highlight = highlight;
+    await foundCard.save();
+
+    res.status(200).json({ message: 'card updated', newCard: foundCard });
 };
 
 const deleteCard = async (req, res) => {
     const { id } = req.params;
-    const removed = await Card.findOneAndDelete({ _id: id });
 
-    if (!removed) {
-        return res.status(404).json({ error: 'Card not found' });
-    }
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
+
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
+
+    await Card.findOneAndDelete({ _id: id });
 
     res.status(200).json({ message: 'Card removed successfully' });
 };
@@ -93,33 +143,20 @@ const copyCard = async (req, res) => {
     const { id } = req.params;
     const { rank } = req.body;
 
-    const foundCard = await Card.findById(id);
+    const foundCard = await cardById(id, { lean: true });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
 
-    if (!foundCard) {
-        return res.status(404).json({ error: 'Card not found' });
-    }
-
-    const {
-        title,
-        description,
-        listId,
-        highlight,
-        owner,
-    } = foundCard;
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, req.user.username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
 
     const newCard = new Card({
-        _id: new mongoose.Types.ObjectId(),
-        order: rank,
-        title,
-        description,
-        listId,
-        owner,
-        highlight,
+        ...foundCard, _id: new mongoose.Types.ObjectId(), rank
     });
 
-    newCard.save();
+    await newCard.save();
 
-    res.status(200).json({ msg: 'Card copied successfully', newCard });
+    res.status(200).json({ msg: 'Card copied successfully', newCard: newCard });
 };
 
 const updateOwner = async (req, res) => {
@@ -127,35 +164,22 @@ const updateOwner = async (req, res) => {
     const { id } = req.params;
     const { ownerName } = req.body;
 
-    const foundUser = await User.findOne({ username }).lean();
-    if (!foundUser) return res.status(403).json({ msg: "User not found" });
+    const foundCard = await cardById(id, { lean: false });
+    if (!foundCard) return res.status(404).json({ error: 'Card not found' });
 
-    //if (!ownerName) {
-    //    const result = await Card.findByIdAndUpdate(
-    //        id, { $set: { owner: null } }, { new: true }
-    //    );
-    //
-    //    return res.status(200).json({ msg: 'Update member successfully', newCard: result });
-    //}
-    //
-    //const foundMember = await User.findOne({ username: ownerName }).lean();
-    //if (!foundMember) return res.status(403).json({ msg: "Member not found" });
-    //
-    //const result = await Card.findByIdAndUpdate(
-    //    id, { $set: { owner: foundMember.username } }, { new: true }
-    //);
+    const { boardId } = foundCard;
+    const { authorized } = await isActionAuthorized(boardId, username);
+    if (!authorized) return res.status(403).json({ msg: 'unauthorized' });
 
-    const result = await Card.findByIdAndUpdate(
-        id, { $set: { owner: ownerName } }, { new: true }
-    );
+    foundCard.owner = ownerName;
+    await foundCard.save();
 
-    res.status(200).json({ msg: 'Update member successfully', newCard: result });
+    res.status(200).json({ msg: 'Update member successfully', newCard: foundCard });
 };
 
 module.exports = {
     getCard,
     addCard,
-    updateCard,
     updateTitle,
     updateDescription,
     updateHighlight,
